UI Update - Re Designed Health Bar, Adjust Lose/Pause Back Images, Removed Win Back Image (will replace!)
Added a new asset to the health bar with a heart and tested it and it does work as previous health bar did. Shrunk the backgrounds for the lose and win menus and move the enemy count and new health bar up a little since I noticed enemy count was in the background and due to the black lettering, it could not be seen. Removed the back image for win but will be updating it with a new image due to previous image not allowing me to adjust it.