ItemPickup Error Resolved
Added a SerializeField for the ItemPickup script. Attached the script to the Player prefab. In the main camera field in Unity, dragged the main camera from the Player prefab so that it recognizes the correct camera when going to pickup items.

All has been done in my scene in Unity.