ItemPickup Issues
Noticed some errors when going to pickup items. Thought trying to use a raycast on the player camera would be a good idea, but became too complicated and removed all code from that. Reviewed pickup code from the professor and seems to work, but there is a conflict with the weapon and weapon stats that causes an error. Commented it out and left a comment note about it for review later today