Moving Platform and Collectible Item
Worked on both making a moving platform and a collectible item. All work has been done in the ZacharyDiston_Scene as to not interfere with any work done in the main scene.
Platform moves back and forth between two waypoints at a slow speed to start, rotating the opposite was once it reaches an endpoint. There is just one here for now as this is just to make sure it works properly.
There is also new scripts relating to these things that I have worked on as well. For the collectible one, it can be assigned to our player in main scene to act as an inventory system for any object the player comes across and can pickup.