Update to Moving Platforms and added One Way Platforms
Wanted to update the moving platforms with different directions for vertical and diagonal movement so we had all the bases covered for whichever direction we want the player to go.
Added one way platforms and code for said platforms and colliders for those specifically. Also added different angles of these platforms as well for a variety to choose from as the player advances through the levels.
Built the game on my own device to grab data for Dez.