Modified enemy ai script
To save time and not repeat code, I decided to use the EnemyAI script for the bosses as well. I have added a check that if they are tagged "Enemy" they will still roam and walk around but if not then they will not. This prevents the boss from wandering during boss fights