Removed True Horror asset and Shadow Boss script
Ran into issue earlier with using the clowns in the True Horror pack so I had to go and use a different one so I removed the asset so it doesn't take up unnecessary space. Same with the Shadow Boss UI I had started. Just decided to use the EnemyAI script with modifications to detect whether it's an enemy or boss